최종 수정 전 마지막 단계

+++ src/main/java/com/Study/vibeclean/dto/stm/request/StmRequest.java
@@ -12,14 +12,14 @@ public class StmRequest {
 
     @Data
     public static class Position {
-        private int x;
-        private int y;
+        private float x;
+        private float y;
 
-        public int getX() {
+        public float getX() {
             return x;
         }
 
-        public int getY() {
+        public float getY() {
             return y;
         }
     }
